添加查询总页数 添加页面页码相关

+++ dogcore/src/main/java/com/bird/dao/BaseDao.java
@@ -40,6 +40,6 @@ public interface BaseDao<T> {
     List<T> findByPage(Object O);
 
     //查询某表的总数
-    long getCount(Object o);
+    int getCount(Object o);
     //--------------------------
 }
